requirements.txt
fastapi
uvicorn
python-dotenv
opencv-python-headless
requests
mysql-connector-python
imagehash